Update formated_address method and moved address comfirmation methods to seprate module to free up the model
